CASE 1 - updatePolicy Success

+++ src/main/java/com/policywebserviceapi/repository/PolicyRepository.java
@@ -4,6 +4,6 @@ import com.policywebserviceapi.model.Policy;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-public interface PolicyRepository extends JpaRepository<Policy, Integer>{
+public interface PolicyRepository extends JpaRepository<Policy, Long>{
 
 }
